+++ packages/design-system/src/components/ErrorState/ErrorState.module.scss
