+++ packages/design-system/src/components/ErrorState/illutstrations/ErrorIllustration.tsx
