--- packages/design-system/src/components/WIP/ErrorState/ErrorState.stories.mdx
@@ -1,118 +1,0 @@
-import { ArgsTable, Meta, Canvas, Story } from '@storybook/addon-docs';
-import { FigmaImage, Use } from '~docs';
-
-import ErrorState from './ErrorState';
-
-import * as Stories from './ErrorState.stories';
-
-<Meta
-	title="[WIP] Components/ErrorState"
-	component={ErrorState}
-	parameters={{
-		status: { figma: 'wip', storybook: 'wip', react: 'wip', i18n: 'wip' },
-		figmaLink: 'https://www.figma.com/file/o5kR1lrKR9he56PyyfWaTa/Error-States',
-	}}
-/>
-
-# ErrorState
-
-Definition
-
-They describe the state where the app fails to complete an action or can't respond or do what the user wants.
-
-The scope of the error varies and could be associated with:
-
-- The entire app
-- The entire current screen
-- A specific element on the screen
-
-Depending on the scope, the error state may display directly on the screen, in dialogs, or in toasters.
-
-## Zoning
-
-<FigmaImage src="" alt="Zoning image for ErrorState" />
-
-## Style
-
-<FigmaImage src="https://www.figma.com/file/o5kR1lrKR9he56PyyfWaTa/Error-States?node-id=104%3A5159" alt="Styles for ErrorStates" />
-
-### Usage
-
-Use this error state to clearly tell users what went wrong, the underlying cause and how they could fix the problem, take a step back, or move to other pages.
-
-This error state includes:
-
-- an icon which supports the error message (except in alert toasters)
-- an informative title to clearly state the problem
-- body text to:
-  - Say why this happened and give any additional information or context users need to know in order to do the right action. Avoid repeating content from the title
-    - suggest actions users can take
-    - give users at least one button (primary CTA) to help them resolve the problem and a secondary outline button, a button with an option to go back and cancel the action. (except in alert toasters)
-
-### Variations
-
-Error state visually varies depending on what features are provided through the props to the component.
-
-#### Default (with a title and a description)
-
-<Canvas>
-	<Story story={Stories.DefaultStory} />
-</Canvas>
-
-#### With a link (default features, plus a link)
-
-<Canvas>
-	<Story story={Stories.WithLink} />
-</Canvas>
-
-#### With a link built from `<Link />` props
-
-<Canvas>
-	<Story story={Stories.WithLinkProps} />
-</Canvas>
-
-#### With an action (default features, plus an action)
-
-<Canvas>
-	<Story story={Stories.WithAction} />
-</Canvas>
-
-#### With an action and a link (default features, plus an action and a link)
-
-<Canvas>
-	<Story story={Stories.WithActionAndLink} />
-</Canvas>
-
-## States
-
-N/A
-
-## Content:
-
-Use error states to help users fix the problem and let the product voice shine through.
-
-### Title
-
-Template: Couldn't (do something)
-
-### Body
-
-Template: (cause) + (actions to do)
-
-### CTA
-
-CTAs which align with the propositions in the body
-
-### Example:
-
-Title: Couldn’t rerun the preparation
-Body: The below parameters have changed in the last run. Edit them and try again, or define a new run.
-CTA: Start over Edit
-
-## Interactions
-
-## Accessibility
-
-N/A
-
-## Usage
