+++ packages/design-system/src/index.ts
@@ -24,7 +24,7 @@ import EmptyState, {
 	EmptyStateMedium,
 	EmptyStateSmall,
 } from './components/EmptyState';
-import ErrorState from './components/WIP/ErrorState';
+import ErrorState from './components/ErrorState';
 import Form from './components/Form';
 import { AffixButton, AffixReadOnly, AffixSelect } from './components/Form/Affix';
 import HeaderBar from './components/WIP/HeaderBar';
